requirements.txt
anthropic
requests
beautifulsoup4
lxml
python-dotenv
arelle-release
